requirements-test.txt
-r requirements.txt
pytest-django
pytest-playwright

